ПЕ-22б Клементьев М.А.: Запрос на создание схемы и таблицы Person.lecturer

+++ SQL/Task_16/create Person.lecturer.sql
@@ -8,7 +8,7 @@ create table Person.lecturer
 	even_date datetime,
 	person tinyint,
 	cathedra tinyint,
-	act int,
+	act tinyint,
 	constraint PK__lecturer primary key (lecturer),
 	constraint FK__lecturer__person foreign key (person) references Person.person,
 	constraint FK__lecturer__cafedra foreign key (cathedra) references Cathedra.cathedra
